cmake now builds botlib (as a shared object) fix q3 looping sounds fix q3sv bug that was kicking clients on map changes attempt to resize q3ui if the window changes size added some more disconnect reasons (for menuqc/q3ui to report). reworked particle count/step arguments for better compat with DP. particles that used count for trails were already broken... drawtextfield builtin will report line numbers shown, so qc can finally tell how much text there actually was added some more items to 'fps_preset dp', stuff needed for the 'quake15' mod. added dpcompat_noretouchground cvar for people wanting to mimic dp's bugs. added 'r_netgraph 2' setting, which will show packet sizes per svc, in part to highlight wasteful mods. added cvar to disable the q1mdl floodfill, which caused problems for yet another person. internal editor now attempts to support debugging asm, if no source is available. fix 64bit icon bug in x11. FINALLY fix high's te_teleport effect. load with no arguments will scan for the most recent a#/s#/quick savedgame instead of using just quick. load command will chose between fte and vanilla savedgame formats based on modification time, if both exist in the same package/dir.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@5394 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/common/vm.h
@@ -67,7 +67,7 @@ qboolean UI_OpenMenu(void);
 void UI_Restart_f(void);
 qboolean UI_Q2LayoutChanged(void);
 void UI_StringChanged(int num);
-qboolean UI_MousePosition(int xpos, int ypos);
+qboolean UI_MousePosition(float xpos, float ypos);
 int UI_MenuState(void);
 qboolean UI_KeyPress(int key, int unicode, qboolean down);
 void UI_Reset(void);
